+++ src/Command/Tuc.fs
@@ -36,13 +36,15 @@ module Tuc =
                     |> List.iter (Dump.parsedTuc output)
 
                     ExitCode.Success
-                | Error error ->
+                | Error errors ->
                     output.Message <| sprintf "\n<c:gray>%s</c>\n" ("-" |> String.replicate 100)
 
-                    error
-                    |> ParseError.format baseIndentation
-                    |> output.Message
-                    |> output.NewLine
+                    errors
+                    |> List.iter (
+                        ParseError.format baseIndentation
+                        >> output.Message
+                        >> output.NewLine
+                    )
 
                     ExitCode.Error
 
@@ -53,9 +55,9 @@ module Tuc =
                 |> List.map (fun tucFile ->
                     match tucFile |> Parser.parse output domainTypes with
                     | Ok tucs -> [ tucFile; (tucs |> List.length |> string); "OK"; "" ]
-                    | Error error ->
+                    | Error errors ->
                         exitCode <- ExitCode.Error
-                        [ tucFile; "0"; "Error"; error |> ParseError.errorName ]
+                        [ tucFile; "0"; "Error"; errors |> List.map (ParseError.errorName) |> List.distinct |> String.concat ", " ]
                 )
                 |> output.Table [ "Tuc file"; "Tucs in file"; "Status"; "Detail" ]
 
@@ -143,14 +145,14 @@ module Tuc =
                 let! tucs =
                     tucFile
                     |> Parser.parse output domainTypes
-                    |> Result.mapError (ParseError.format baseIndentation)
+                    |> Result.mapError (List.map (ParseError.format baseIndentation))
 
                 let! tucs =
                     match specificTuc with
                     | Some specific ->
                         tucs
                         |> List.tryFind (Tuc.name >> (=) (TucName specific))
-                        |> Result.ofOption (sprintf "<c:red>Specific tuc %A is not parsed.</c>" specificTuc)
+                        |> Result.ofOption [ sprintf "<c:red>Specific tuc %A is not parsed.</c>" specificTuc ]
                         |> Result.map List.singleton
                     | _ -> Ok tucs
 
@@ -161,6 +163,7 @@ module Tuc =
                     tucs
                     |> Generate.puml output pumlName
                     |> Result.mapError PumlError.format
+                    |> Validation.ofResult
 
                 match outputFile with
                 | Some outputFile ->
@@ -200,10 +203,9 @@ module Tuc =
                 | Ok message ->
                     output.Success message
                     ExitCode.Success
-                | Error message ->
-                    message
-                    |> output.Message
-                    |> output.NewLine
+                | Error messages ->
+                    messages
+                    |> List.iter (output.Message >> output.NewLine)
                     ExitCode.Error
 
         match input with