+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add domain type for `Handler`
 - Allow to use FQ name for events and use deeper event name
 - Allow to check multiple .tuc files in a dir at once
+- Parse .tuc with more errors, if possible
 
 ## 0.1.0 - 2020-08-12
 - Initial implementation